change UserId from int to long

+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Interfaces/IUploadFileService.cs
@@ -2,6 +2,6 @@
 {
     public interface IUploadFileService
     {
-        public Task<string> UploadFileBlobAsync(string filePath, string fileName, int userId);
+        public Task<string> UploadFileBlobAsync(string filePath, string fileName, long userId);
     }
 }
